Initial state for What's New in Spring Authorization Server 1.3
This commit demonstrates a scenario for OAuth2 that combines both
authentication and authorization into a single configuration.

The application uses a single ClientRegistration "oidc-client" to
request the following scopes:

- openid
- profile
- cashcard:read
- cashcard:write
- purchase:read

With this configuration, we can use a single access token obtained at
login to request resources from the following resource servers:

- cashcard-service - an API to manage cashcards with the ability to
  retrieve cards (requires cashcard:read scope) and add new ones
  (requires cashcard:write scope)

- purchase-service - an API to retrieve purchases for a cashcard
  (requires purchase:read scope)

Note that the user is only required to provide consent once during login
and the access token can be re-used for multiple resource servers.