Add data-service and token-exchange
This commit demonstrates splitting a "monolith" into one or more
microservices and using OAuth 2.0 Token Exchange to propagate the user
from an incoming access token into an outgoing access token.

Note that the `aud` claim of the token is validated on each server,
so that tokens are only valid for a particular resource server.