Add token-exchange to gateway
This commit demonstrates bypassing the "monolith" by directly using
token-exchange from the gateway to obtain an access token with the
subject of the current user. The idea here is to incrementally improve
the architecture by:

1. Creating a new microservice called by the monolith but leaving the
   gateway unchanged.
2. Eventually changing the gateway to bypass the monolith and call the
   microservice directly.

Note that the `aud` claim of the token obtained by the gateway is still
correct for the new microservice.

Note also that with this change, the gateway no longer requests scopes
other than during login. Use of authorization grants become an internal
implementation detail hidden from the user.