Enable Token Exchange on the client and server
Authorization Server:

This commit enables OAuth2 Clients (cashcard-client and purchase-client)
to use the Token Exchange grant, which is supported out of the box in
Spring Authorization Server 1.3.

No other configuration is required on the auth server. Applications can
configure additional customizations in order to further restrict the
authorization of such clients. For example, optional parameters such
as audience and resource can be used on the server to provide custom
decisioning, enforce policies, etc. See
  https://spring.io/blog/2024/03/19/token-exchange-support-in-spring-security-6-3-0-m3
for an example.

Client:

This commit configures support for Token Exchange by registering a bean
of type TokenExchangeReactiveOAuth2AuthorizedClientProvider, which is
automatically picked up by Spring Security and wired into a default
instance of ReactiveOAuth2AuthorizedClientManager.

Additionally, the configuration demonstrates setting up a custom
subjectTokenResolver that resolves the access token obtained at login,
which is used to perform token exchange.

By doing so, the user (`sub` claim of the access token) is preserved
during the token exchange such that the new access token also belongs
to the user. However, new scopes are requested by the client which are
internal to the system, and don't require user consent. This means that
each individual resource server the client is calling receives an access
token meant for it alone, with exactly the scope(s) necessary and no
unnecessary scopes, which reduces the attack surface of a leaked token.
The user experience is also improved because the user does not need to
consent to internal scopes being requested by the client, yet security
of the overall system is preserved.

Note that the specific token exchange flow being used here is referred
to as "impersonation", where the client is granted permission by the
authorization server to act on behalf of the user with scopes that
are understood by the resource server. The user does not grant consent
for the scopes because they are internal to the system. In some cases,
such consent may be provided separately (out of band), such as through
an administrative console for example. In this scenario, we imagine
that the grant is between the client and the authorization server only,
and the user is unaware that token exchange is being performed or even
that OAuth2 is being used at all.